+++ components/composite/StepPlaceOrder/index.tsx
@@ -1,5 +1,3 @@
-import { PlaceOrderContainer } from "@commercelayer/react-components"
-import { useRouter } from "next/router"
 import { useContext, useState } from "react"
 import { Trans, useTranslation } from "react-i18next"
 
@@ -34,7 +32,6 @@ const StepPlaceOrder: React.FC<Props> = ({
   privacyUrl,
 }) => {
   const { t } = useTranslation()
-  const { query } = useRouter()
 
   const [isPlacingOrder, setIsPlacingOrder] = useState(false)
 
@@ -43,16 +40,6 @@ const StepPlaceOrder: React.FC<Props> = ({
 
   if (!appCtx) {
     return null
-  }
-  let paypalPayerId = ""
-  let checkoutComSession = ""
-
-  if (query.PayerID) {
-    paypalPayerId = query.PayerID as string
-  }
-
-  if (query["cko-session-id"]) {
-    checkoutComSession = query["cko-session-id"] as string
   }
 
   const { placeOrder } = appCtx
@@ -102,51 +89,45 @@ const StepPlaceOrder: React.FC<Props> = ({
           }}
         </StyledErrors>
       </ErrorsContainer>
-      <PlaceOrderContainer
-        options={{
-          paypalPayerId,
-          checkoutCom: { session_id: checkoutComSession },
-        }}
-      >
-        <>
-          {!!termsUrl && !!privacyUrl && (
-            <FlexContainer className="items-start mx-5 mt-4 mb-2.5 md:mb-5 md:pb-5 md:mx-0 md:mt-0 md:border-b lg:pl-8">
-              <StyledPrivacyAndTermsCheckbox
-                id="privacy-terms"
-                className="relative form-checkbox top-0.5"
-                data-test-id="checkbox-privacy-and-terms"
-              />
-              <Label htmlFor="privacy-terms">
-                <Trans
-                  i18nKey="general.privacy_and_terms"
-                  components={{
-                    bold: <strong />,
-                    termsUrl: (
-                      <a href={termsUrl} target="_blank" rel="noreferrer" />
-                    ),
-                    privacyUrl: (
-                      <a href={privacyUrl} target="_blank" rel="noreferrer" />
-                    ),
-                  }}
-                />
-              </Label>
-            </FlexContainer>
-          )}
-          <PlaceOrderButtonWrapper>
-            <StyledPlaceOrderButton
-              data-test-id="save-payment-button"
-              isActive={isActive}
-              onClick={handlePlaceOrder}
-              label={
-                <>
-                  {isPlacingOrder && <SpinnerIcon />}
-                  {t("stepPayment.submit")}
-                </>
-              }
+
+      <>
+        {!!termsUrl && !!privacyUrl && (
+          <FlexContainer className="items-start mx-5 mt-4 mb-2.5 md:mb-5 md:pb-5 md:mx-0 md:mt-0 md:border-b lg:pl-8">
+            <StyledPrivacyAndTermsCheckbox
+              id="privacy-terms"
+              className="relative form-checkbox top-0.5"
+              data-test-id="checkbox-privacy-and-terms"
             />
-          </PlaceOrderButtonWrapper>
-        </>
-      </PlaceOrderContainer>
+            <Label htmlFor="privacy-terms">
+              <Trans
+                i18nKey="general.privacy_and_terms"
+                components={{
+                  bold: <strong />,
+                  termsUrl: (
+                    <a href={termsUrl} target="_blank" rel="noreferrer" />
+                  ),
+                  privacyUrl: (
+                    <a href={privacyUrl} target="_blank" rel="noreferrer" />
+                  ),
+                }}
+              />
+            </Label>
+          </FlexContainer>
+        )}
+        <PlaceOrderButtonWrapper>
+          <StyledPlaceOrderButton
+            data-test-id="save-payment-button"
+            isActive={isActive}
+            onClick={handlePlaceOrder}
+            label={
+              <>
+                {isPlacingOrder && <SpinnerIcon />}
+                {t("stepPayment.submit")}
+              </>
+            }
+          />
+        </PlaceOrderButtonWrapper>
+      </>
     </>
   )
 }