+++ specs/e2e/GTM.spec.ts
@@ -3,6 +3,8 @@ import { faker } from "@faker-js/faker"
 import { test, expect } from "../fixtures/tokenizedPage"
 import { euAddress, usAddress } from "../utils/addresses"
 
+const TIMEOUT = 1000
+
 test.describe("multi shipments", () => {
   const customerEmail = faker.internet.email().toLocaleLowerCase()
 
@@ -15,7 +17,7 @@ test.describe("multi shipments", () => {
         privacy_url: "https://www.google.it",
       },
       organization: {
-        gtmId: "GTM-123456",
+        gtm_id_test: "GTM-123456",
       },
       lineItemsAttributes: [
         { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
@@ -35,14 +37,22 @@ test.describe("multi shipments", () => {
     expect(dataLayer[0].ecommerce.currency).toBe("EUR")
     expect(dataLayer[0].ecommerce.value).toBe(244)
     expect(dataLayer[0].ecommerce.items?.length).toBe(2)
+    expect(
+      dataLayer[0].ecommerce.items?.filter(
+        (item) => item?.item_id === "CANVASAU000000FFFFFF1824"
+      )
+    ).toHaveLength(1)
 
     await checkoutPage.checkStep("Shipping", "open")
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
 
     await checkoutPage.selectShippingMethod({ text: "Standard Shipping" })
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
     await checkoutPage.selectShippingMethod({
       text: "Standard Shipping",
       shipment: 1,
     })
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
 
     await checkoutPage.save("Shipping")
 
@@ -60,13 +70,17 @@ test.describe("multi shipments", () => {
 
     await checkoutPage.clickStep("Shipping")
 
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
     await checkoutPage.selectShippingMethod({ text: "Express Delivery" })
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
     await checkoutPage.selectShippingMethod({
       text: "Express Delivery",
       shipment: 1,
     })
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
 
     await checkoutPage.save("Shipping")
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
 
     dataLayer = await checkoutPage.getDataLayer("add_shipping_info")
     expect(dataLayer.length).toBe(4)
@@ -81,15 +95,20 @@ test.describe("multi shipments", () => {
     expect(dataLayer[3].ecommerce.items?.length).toBe(1)
 
     await checkoutPage.clickStep("Shipping")
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
 
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
     await checkoutPage.selectShippingMethod({ text: "Express Delivery" })
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
     await checkoutPage.selectShippingMethod({
       text: "Standard Shipping",
       shipment: 1,
     })
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
 
     await checkoutPage.save("Shipping")
 
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
     dataLayer = await checkoutPage.getDataLayer("add_shipping_info")
     expect(dataLayer.length).toBe(6)
     expect(dataLayer[4].ecommerce.currency).toBe("EUR")
@@ -103,15 +122,19 @@ test.describe("multi shipments", () => {
     expect(dataLayer[5].ecommerce.items?.length).toBe(1)
 
     await checkoutPage.clickStep("Shipping")
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
 
     await checkoutPage.selectShippingMethod({ text: "Standard Shipping" })
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
     await checkoutPage.selectShippingMethod({
       text: "Express Delivery",
       shipment: 1,
     })
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
 
     await checkoutPage.save("Shipping")
 
+    await checkoutPage.page.waitForTimeout(TIMEOUT)
     dataLayer = await checkoutPage.getDataLayer("add_shipping_info")
     expect(dataLayer.length).toBe(8)
     expect(dataLayer[6].ecommerce.currency).toBe("EUR")
@@ -145,7 +168,7 @@ test.describe("single shipment", () => {
         customer_email: customerEmail,
       },
       organization: {
-        gtmId: "GTM-123456",
+        gtm_id_test: "GTM-123456",
       },
       lineItemsAttributes: [
         { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
@@ -214,10 +237,10 @@ test.describe("with single shipping method", () => {
   test.use({
     defaultParams: {
       organization: {
-        gtmId: "GTM-123456",
+        gtm_id_test: "GTM-123456",
       },
       order: "with-items",
-      market: process.env.NEXT_PUBLIC_MARKET_ID_SINGLE_SHIPPING_METHOD,
+      market: process.env.E2E_MARKET_ID_SINGLE_SHIPPING_METHOD,
       lineItemsAttributes: [
         { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
       ],
@@ -283,6 +306,140 @@ test.describe("with single shipping method", () => {
     expect(dataLayer[0].ecommerce.currency).toBe("USD")
     expect(dataLayer[0].ecommerce.shipping_tier).toBe("Express Delivery")
     expect(dataLayer[0].ecommerce.value).toBe(7)
+    expect(dataLayer[0].ecommerce.items?.length).toBe(1)
+  })
+})
+
+test.describe("with digital products", () => {
+  const customerEmail = faker.internet.email().toLocaleLowerCase()
+
+  test.use({
+    defaultParams: {
+      organization: {
+        gtm_id_test: "GTM-123456",
+      },
+      order: "digital",
+      market: process.env.E2E_MARKET_ID_SINGLE_SHIPPING_METHOD,
+      lineItemsAttributes: [
+        { sku_code: "CANVASAU000000FFFFFF1824", quantity: 1 },
+      ],
+      orderAttributes: {
+        customer_email: customerEmail,
+      },
+    },
+  })
+
+  test("do not fire add_shipping_info event", async ({ checkoutPage }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+
+    await checkoutPage.setBillingAddress(usAddress)
+    await checkoutPage.save("Customer")
+
+    let dataLayer = await checkoutPage.getDataLayer("begin_checkout")
+    expect(dataLayer.length).toBe(1)
+    expect(dataLayer[0].ecommerce.currency).toBe("USD")
+    expect(dataLayer[0].ecommerce.value).toBe(15)
+    expect(dataLayer[0].ecommerce.items?.length).toBe(1)
+
+    await checkoutPage.checkStep("Customer", "close")
+    await checkoutPage.checkStep("Shipping", "not_present")
+    await checkoutPage.checkStep("Payment", "open")
+
+    dataLayer = await checkoutPage.getDataLayer("add_shipping_info")
+    expect(dataLayer.length).toBe(0)
+
+    await checkoutPage.selectPayment("wire")
+
+    await checkoutPage.save("Payment")
+
+    dataLayer = await checkoutPage.getDataLayer("add_payment_info")
+    expect(dataLayer.length).toBe(1)
+    expect(dataLayer[0].ecommerce.currency).toBe("USD")
+    expect(dataLayer[0].ecommerce.payment_type).toBe("Wire Transfer")
+    expect(dataLayer[0].ecommerce.value).toBe(0)
+    expect(dataLayer[0].ecommerce.items?.length).toBe(1)
+
+    dataLayer = await checkoutPage.getDataLayer("purchase")
+    expect(dataLayer.length).toBe(1)
+    expect(dataLayer[0].ecommerce.currency).toBe("USD")
+    expect(dataLayer[0].ecommerce.value).toBe(18.3)
+    expect(dataLayer[0].ecommerce.shipping).toBe(0)
+    expect(dataLayer[0].ecommerce.tax).toBe(3.3)
+  })
+})
+
+test.describe("with bundle", () => {
+  const customerEmail = faker.internet.email().toLocaleLowerCase()
+
+  test.use({
+    defaultParams: {
+      order: "with-items",
+      orderAttributes: {
+        customer_email: customerEmail,
+      },
+      organization: {
+        gtm_id_test: "GTM-123456",
+      },
+      lineItemsAttributes: [{ bundle_code: "SHIRTSETSINGLE", quantity: 1 }],
+      addresses: {
+        billingAddress: euAddress,
+        sameShippingAddress: true,
+      },
+    },
+  })
+
+  test("check payment and purchase", async ({ checkoutPage }) => {
+    await checkoutPage.checkOrderSummary("Order Summary")
+    let dataLayer = await checkoutPage.getDataLayer("begin_checkout")
+    expect(dataLayer.length).toBe(1)
+    expect(dataLayer[0].ecommerce.currency).toBe("EUR")
+    expect(dataLayer[0].ecommerce.value).toBe(105)
+    expect(
+      dataLayer[0].ecommerce.items?.filter(
+        (item) => item?.item_id === "SHIRTSETSINGLE"
+      )
+    ).toHaveLength(1)
+
+    expect(dataLayer[0].ecommerce.items?.length).toBe(1)
+
+    await checkoutPage.checkStep("Shipping", "open")
+
+    await checkoutPage.selectShippingMethod({ text: "Standard Shipping" })
+
+    await checkoutPage.save("Shipping")
+
+    dataLayer = await checkoutPage.getDataLayer("add_shipping_info")
+    expect(dataLayer.length).toBe(1)
+    expect(dataLayer[0].ecommerce.currency).toBe("EUR")
+    expect(dataLayer[0].ecommerce.shipping_tier).toBe("Standard Shipping")
+    expect(dataLayer[0].ecommerce.value).toBe(0)
+    expect(dataLayer[0].ecommerce.items?.length).toBe(2)
+
+    await checkoutPage.selectPayment("stripe")
+
+    await checkoutPage.setPayment("stripe")
+
+    await checkoutPage.save("Payment")
+
+    dataLayer = await checkoutPage.getDataLayer("add_payment_info")
+    expect(dataLayer.length).toBe(1)
+    expect(dataLayer[0].ecommerce.currency).toBe("EUR")
+    expect(dataLayer[0].ecommerce.payment_type).toBe("Stripe Payment")
+    expect(dataLayer[0].ecommerce.value).toBe(10)
+    expect(dataLayer[0].ecommerce.items?.length).toBe(1)
+
+    dataLayer = await checkoutPage.getDataLayer("purchase")
+    expect(dataLayer.length).toBe(1)
+    expect(dataLayer[0].ecommerce.currency).toBe("EUR")
+    expect(dataLayer[0].ecommerce.value).toBe(115)
+    expect(dataLayer[0].ecommerce.shipping).toBe(0)
+    expect(dataLayer[0].ecommerce.tax).toBe(0)
+
+    const orderNumber = await checkoutPage.getOrderNumber()
+
+    expect(dataLayer[0].ecommerce.transaction_id).toBe(
+      parseInt(orderNumber.replace("#", ""))
+    )
     expect(dataLayer[0].ecommerce.items?.length).toBe(1)
   })
 })