+++ utils/isHosted.ts
@@ -1,0 +1,3 @@
+export function isHosted() {
+  return !!process.env.NEXT_PUBLIC_HOSTED
+}